+++ packages/solana/vue/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @axiajs/solana.vue
 
+## 0.1.0-alpha.2
+
+### Minor Changes
+
+- Add fetch user NFTs composable
+
+### Patch Changes
+
+- Updated dependencies
+  - @axiajs/solana.core@0.1.0-alpha.0
+
 ## 0.1.0-alpha.1
 
 ### Minor Changes